Move <NotificationWarning> to its own module
