+++ src/components/Layout.js
@@ -1,128 +1,17 @@
 import React, { useState, useContext } from 'react';
-import Sidebar from './Sidebar';
 import { AiOutlineMenu } from "react-icons/ai";
-import { FaExclamationTriangle, FaTimes } from 'react-icons/fa';
-import logo from '../assets/images/wallet_white.png';
 import { useLocation, useNavigate } from 'react-router-dom';
+import { useTranslation } from 'react-i18next';
+
+import Sidebar from './Sidebar';
+import logo from '../assets/images/wallet_white.png';
 import WelcomeTourGuide from './WelcomeTourGuide/WelcomeTourGuide';
-import { useSessionStorage } from '../components/useStorage';
-import { Trans, useTranslation } from 'react-i18next';
 import BottomNav from './BottomNav';
 import OnlineStatusContext from '../context/OnlineStatusContext';
-import { notificationApiIsSupported } from '../firebase';
 import { PiWifiHighBold, PiWifiSlashBold } from "react-icons/pi";
-import SessionContext from '../context/SessionContext';
 import FadeInContentTransition from './FadeInContentTransition';
+import NotificationWarning from '../components/NotificationWarning';
 
-
-const NotificationWarning = ({ isPermissionGranted, tokenSentInSession }) => {
-	const { isOnline } = useContext(OnlineStatusContext);
-	const { api } = useContext(SessionContext);
-	const [isMessageNoGrantedVisible, setIsMessageNoGrantedVisible,] = api.useClearOnClearSession(useSessionStorage('isMessageNoGrantedVisible', false));
-	const [isMessageGrantedVisible, setIsMessageGrantedVisible,] = api.useClearOnClearSession(useSessionStorage('isMessageGrantedVisible', false));
-	const [isMessageOfflineVisible, setIsMessageOfflineVisible,] = api.useClearOnClearSession(useSessionStorage('isMessageOfflineVisible', false));
-
-	const handleCloseMessageOffline = () => {
-		setIsMessageOfflineVisible(true);
-	};
-
-	const handleCloseMessageNoGranted = () => {
-		setIsMessageNoGrantedVisible(true);
-	};
-
-	const handleCloseMessageGranted = () => {
-		setIsMessageGrantedVisible(true);
-	};
-
-	const show = (
-		(isOnline === false && isMessageOfflineVisible === false)
-		|| (
-			isOnline === true
-			&& isPermissionGranted != null
-			&& notificationApiIsSupported()
-			&& (
-				(!isPermissionGranted && isMessageNoGrantedVisible === false)
-				|| (
-					isPermissionGranted
-					&& tokenSentInSession === false
-					&& isMessageGrantedVisible === false)
-			))
-	);
-
-	return (
-		show
-			? (
-				<div className="bg-orange-100 shadow-lg p-4 rounded-lg mb-4 flex items-center">
-					<div className="mr-4 text-orange-500">
-						<FaExclamationTriangle size={24} />
-					</div>
-
-					{isOnline === false && isMessageOfflineVisible === false && (
-						<>
-							<div className="flex-grow">
-								<p className='text-sm'>
-									<Trans
-										i18nKey="layout.messageOffline"
-										components={{ strong: <strong /> }}
-									/>
-								</p>
-							</div>
-							<button
-								className="ml-2 text-gray-800"
-								onClick={handleCloseMessageOffline}
-							>
-								<FaTimes size={24} />
-							</button>
-						</>
-					)}
-					{isOnline === true && (
-						<>
-							{!isPermissionGranted && (
-								<>
-									<div className="flex-grow">
-										<p className='text-sm'>
-											<Trans
-												i18nKey="layout.messageAllowPermission"
-												components={{ strong: <strong /> }}
-											/>
-										</p>
-									</div>
-									<button
-										className="ml-2 text-gray-800"
-										onClick={handleCloseMessageNoGranted}
-									>
-										<FaTimes size={24} />
-									</button>
-								</>
-							)}
-							{isPermissionGranted && tokenSentInSession === false && (
-								<>
-									<div className="flex-grow">
-										<p className='text-sm'>
-											<Trans
-												i18nKey="layout.messageResetPermission"
-												components={{
-													strong: <strong />,
-													reloadButton: <button className='text-primary underline' onClick={() => window.location.reload()} />,
-												}}
-											/>
-										</p>
-									</div>
-									<button
-										className="ml-2 text-gray-800"
-										onClick={handleCloseMessageGranted}
-									>
-										<FaTimes size={24} />
-									</button>
-								</>
-							)}
-						</>
-					)}
-				</div>
-			)
-			: <></>
-	);
-};
 
 const Layout = ({ children, isPermissionGranted, tokenSentInSession }) => {
 	const { isOnline } = useContext(OnlineStatusContext);
